+++ controller/src/main/java/com/linbit/linstor/core/apicallhandler/controller/CtrlConfApiCallHandler.java
@@ -991,6 +991,8 @@ public class CtrlConfApiCallHandler
                         case ZfsRollbackStrategy.FULL_KEY_USE_ZFS_ROLLBACK_PROP:
                             // fall-through
                         case ZfsDeleteStrategy.FULL_KEY_ZFS_DELETE_STRATEGY:
+                            // fall-through
+                        case ApiConsts.NAMESPC_DRBD_PROXY + "/" + ApiConsts.KEY_DRBD_PROXY_AUTO_ENABLE:
                             // no need to update stlts
                             setCtrlProp(peerAccCtx.get(), key, normalized, namespace, propChangedListener);
                             break;